requirements.txt
# Core FastAPI and async dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database and async drivers
motor==3.3.2
pymongo==4.6.0

# Task queue and caching
celery==5.3.4
redis==5.0.1

# Web scraping and text extraction
newspaper3k==0.2.8
readability-lxml==0.8.1
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
aiohttp==3.9.1

# OpenAI integration
openai==1.3.7

# WebSocket support
websockets==12.0
python-socketio==5.10.0

# Monitoring and logging
structlog==23.2.0
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
